// backend/model/nota.ts
export interface Nota {
    numNota: string;
    dtAplicacao: string;
    dtVencimento: string;
    valorBase: string;
    taxaAtual: string;
    taxaFinal: string;
    rendBrutoAcum: string;
    pgtRendBrutoAcum: string;
    provisaoIR: string;
    provisaoIOF: string;
    rendLiquidAcum: string;
    pgtRendLiquidAcum: string;
    dtSaldo1: string;
    dtSaldo2: string;
}